Refactor witgen passes
