+++ include/llzk/Transforms/LLZKTransformationPasses.h
@@ -38,12 +38,6 @@ std::unique_ptr<mlir::Pass> createPolyLoweringPass(unsigned maxDegree);
 
 std::unique_ptr<mlir::Pass> createInlineStructsPass();
 
-std::unique_ptr<mlir::Pass> createLowerComputeToCorePass();
-
-std::unique_ptr<mlir::Pass> createCreateWitgenEntryPass();
-
-void addWitgenPreparePipeline(mlir::OpPassManager &pm);
-
 void registerTransformationPassPipelines();
 
 void registerInliningExtensions(mlir::DialectRegistry &registry);
