+++ lib/Transforms/LLZKTransformationPassPipelines.cpp
@@ -15,7 +15,6 @@
 #include "llzk/Transforms/LLZKTransformationPasses.h"
 #include "llzk/Dialect/Polymorphic/Transforms/TransformationPasses.h"
 
-#include <mlir/Conversion/AffineToStandard/AffineToStandard.h>
 #include <mlir/Pass/PassManager.h>
 #include <mlir/Pass/PassRegistry.h>
 #include <mlir/Transforms/Passes.h>
@@ -35,15 +34,6 @@ void addRemoveUnnecessaryOpsAndDefsPipeline(OpPassManager &pm) {
   pm.addPass(llzk::createRedundantReadAndWriteEliminationPass());
   pm.addPass(llzk::createRedundantOperationEliminationPass());
   pm.addPass(llzk::createUnusedDeclarationEliminationPass());
-}
-
-void addWitgenPreparePipeline(OpPassManager &pm) {
-  llzk::polymorphic::FlatteningPassOptions options = {.cleanupMode = llzk::polymorphic::StructCleanupMode::ConcreteAsRoot};
-  pm.addPass(llzk::polymorphic::createFlatteningPass(std::move(options)));
-  pm.addPass(mlir::createLowerAffinePass());
-  pm.addPass(llzk::createInlineStructsPass());
-  pm.addPass(mlir::createCanonicalizerPass());
-  pm.addPass(mlir::createCSEPass());
 }
 
 void registerTransformationPassPipelines() {
@@ -84,11 +74,6 @@ void registerTransformationPassPipelines() {
   }
   );
 
-  PassPipelineRegistration<>(
-      "llzk-witgen-prepare",
-      "Prepare LLZK witness-generation IR for interpreter or execution-engine execution",
-      [](OpPassManager &pm) { addWitgenPreparePipeline(pm); }
-  );
 }
 
 } // namespace llzk
